+++ demisto_sdk/commands/common/errors.py
@@ -87,6 +87,7 @@ ERROR_CODE = {
     "non_existing_docker": "DO107",
     "id_set_conflicts": "ID100",
     "duplicated_id": "ID102",
+    "no_id_set_file": "ID103",
     "remove_field_from_dashboard": "DA100",
     "include_field_in_dashboard": "DA101",
     "remove_field_from_widget": "WD100",
@@ -619,6 +620,11 @@ class Errors:
     def duplicated_id(obj_id):
         return f"The ID {obj_id} already exists, please update the file or update the " \
                f"id_set.json toversion field of this id to match the old occurrence of this id"
+
+    @staticmethod
+    @error_code_decorator
+    def no_id_set_file():
+        return "Unable to find id_set.json file in path - rerun the command with --create-id-set flag"
 
     @staticmethod
     @error_code_decorator